// client/src/components/Clusters/state/types.ts
// Copyright 2019 Authors of Hubble
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
import { Cluster, DiscoverClusterResult } from "../../../graphqlTypes";

export interface ClusterWithSelected {
  readonly selected: boolean;
  readonly cluster: Cluster;
}

export interface State {
  readonly fetching: boolean;
  readonly clusters: Cluster[];
  readonly fetchedOkAtLeastOnce: boolean;
  readonly discoveryResult: DiscoverClusterResult | null;
  readonly discovering: boolean;
}